+++ src/pymatgen/core/surface.py
@@ -507,8 +507,8 @@ class Slab(Structure):
 
         # Determine what fraction the slab is of the total cell size in the
         # c direction. Round to nearest rational number.
-        n_layers_total = int(round(self.lattice.c / self.oriented_unit_cell.lattice.c))
-        n_layers_slab = int(round((sorted_csites[-1].c - sorted_csites[0].c) * n_layers_total))
+        n_layers_total = round(self.lattice.c / self.oriented_unit_cell.lattice.c)
+        n_layers_slab = round((sorted_csites[-1].c - sorted_csites[0].c) * n_layers_total)
         slab_ratio = n_layers_slab / n_layers_total
 
         spg_analyzer = SpacegroupAnalyzer(self)
@@ -1003,8 +1003,8 @@ class SlabGenerator:
                 lcm_miller = lcm(*(miller_index[i] for i, _d in non_orth_ind))
                 for (ii, _di), (jj, _dj) in itertools.combinations(non_orth_ind, 2):
                     scale_factor = [0, 0, 0]
-                    scale_factor[ii] = -int(round(lcm_miller / miller_index[ii]))
-                    scale_factor[jj] = int(round(lcm_miller / miller_index[jj]))
+                    scale_factor[ii] = -round(lcm_miller / miller_index[ii])
+                    scale_factor[jj] = round(lcm_miller / miller_index[jj])
                     slab_scale_factor.append(scale_factor)
                     if len(slab_scale_factor) == 2:
                         break
